requirements.txt
# Framework web
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Base de datos
PyMySQL==1.1.0
cryptography==41.0.7

# Utilidades
python-dotenv==1.0.0
Werkzeug==3.0.1

# Para importación de Excel
openpyxl==3.1.2
pandas==2.1.4

# Validaciones
email-validator==2.1.0

# Fechas
python-dateutil==2.8.2
